fs.promises: reduce readFile memory consumption
fs.promises.readFile used to always take twice the file size's memory,
because it would build buffers, then concatenate them.

This new code copies what fs.readFile does: if the size is known, it
allocates the buffer beforehand. If the size is unknown, buffers are
built in an array, same as the old code.